+++ examples/photos.js
@@ -9,10 +9,9 @@ const createPhotoPiles = async element => {
 
   const piling = createPilingJs(element);
 
-  piling.set('itemSize', 300);
+  piling.set('itemSize', 160);
   piling.set('renderer', imageRenderer);
   piling.set('items', data);
-  piling.set('itemPadding', 10);
 
   piling.set('pileCellAlign', 'center');